Refactor frontend file names
