+++ packages/frontend-app/src/app/pages/Community/ConvertPlaylist.tsx
@@ -34,7 +34,7 @@ import { loadData } from "./loadData";
 import { useApi } from "../../context/ApiContext";
 import useParsedQueryParams from "../../hooks/useParsedQueryParams";
 import { useToast } from "../../components/Toast/ToastContext";
-import * as formValidation from "../../../utils/formValidation";
+import * as formValidation from "../../../utils/form-validation";
 import { constructURL } from "../../../utils/url";
 import { routeIds } from "../../routes";
 import { useNavigate } from "react-router-dom";
