+++ packages/frontend-app/src/app/routes.ts
@@ -1,7 +1,7 @@
 import loadable from "@loadable/component";
 
 import type { RouteObjectWithLoadData } from "react-router-dom";
-import type { ILoadableComponentProps } from "../utils/routeUtils";
+import type { ILoadableComponentProps } from "../utils/route-utils";
 
 import {
   loadData as homeLoadData,
